requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==4.2.7
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
executing==2.0.1
fake-useragent==1.5.1
frozenlist==1.4.1
idna==3.6
ipython==8.21.0
jedi==0.19.1
lxml==5.1.0
matplotlib-inline==0.1.6
multidict==6.0.5
numpy==1.26.4
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
